fixed runtime round by 49days(DWORD -> LONGLONG)
git-svn-id: https://svn.osdn.net/svnroot/switch/switch/trunk@770 6544d933-5e26-41f1-b2f0-1bd7b5b7622e

+++ src/svvitch/Renderer.h
@@ -233,7 +233,7 @@ public:
 	/**
 	 * Scene‚ðƒŒƒ“ƒ_ƒŠƒ“ƒO‚µ‚Ü‚·
 	 */
-	void renderScene(const bool visibled, const DWORD current);
+	void renderScene(const bool visibled, const LONGLONG current);
 
 	const UINT getDisplayAdapters() const;
